Rename internal app.py to onboarding_app.py
