+++ pt_os_web_portal/miniscreen_onboarding/__init__.py
@@ -1,1 +1,1 @@
-from .app import OnboardingApp
+from .onboarding_app import OnboardingApp
